LPS-57384 Simplify API step three -> Create a UniqueFileNames upload hadler specific base class. This class should be used by the apps wich should create a safe file name (with the parenthetical suffix if necessary).
