+++ modules/apps/blogs/blogs-web/src/com/liferay/blogs/web/upload/BaseBlogsImageUploadHandler.java
@@ -19,15 +19,10 @@ import com.liferay.portal.kernel.exception.SystemException;
 import com.liferay.portal.kernel.json.JSONFactoryUtil;
 import com.liferay.portal.kernel.json.JSONObject;
 import com.liferay.portal.kernel.portlet.JSONPortletResponseUtil;
-import com.liferay.portal.kernel.repository.model.FileEntry;
-import com.liferay.portal.kernel.repository.model.Folder;
 import com.liferay.portal.kernel.servlet.ServletResponseConstants;
-import com.liferay.portal.kernel.upload.BaseUploadHandler;
-import com.liferay.portal.kernel.upload.UploadPortletRequest;
+import com.liferay.portal.kernel.upload.BaseUniqueFileNameUploadHandler;
 import com.liferay.portal.kernel.util.FileUtil;
-import com.liferay.portal.kernel.util.ParamUtil;
 import com.liferay.portal.kernel.util.PropsKeys;
-import com.liferay.portal.kernel.util.StreamUtil;
 import com.liferay.portal.kernel.util.StringPool;
 import com.liferay.portal.kernel.util.WebKeys;
 import com.liferay.portal.security.auth.PrincipalException;
@@ -35,17 +30,14 @@ import com.liferay.portal.security.permission.ActionKeys;
 import com.liferay.portal.security.permission.PermissionChecker;
 import com.liferay.portal.security.permission.ResourcePermissionCheckerUtil;
 import com.liferay.portal.theme.ThemeDisplay;
-import com.liferay.portal.util.PortalUtil;
 import com.liferay.portal.util.PrefsPropsUtil;
 import com.liferay.portlet.blogs.BlogImageNameException;
 import com.liferay.portlet.blogs.BlogImageSizeException;
-import com.liferay.portlet.blogs.service.BlogsEntryLocalServiceUtil;
 import com.liferay.portlet.blogs.service.permission.BlogsPermission;
 import com.liferay.portlet.documentlibrary.FileNameException;
 import com.liferay.portlet.documentlibrary.antivirus.AntivirusScannerException;
 
 import java.io.IOException;
-import java.io.InputStream;
 
 import javax.portlet.PortletRequest;
 import javax.portlet.PortletResponse;
@@ -54,49 +46,8 @@ import javax.portlet.PortletResponse;
  * @author Sergio González
  * @author Adolfo Pérez
  */
-public abstract class BaseBlogsImageUploadHandler extends BaseUploadHandler {
-
-	protected FileEntry addFileEntry(PortletRequest portletRequest)
-		throws PortalException {
-
-		UploadPortletRequest uploadPortletRequest =
-			PortalUtil.getUploadPortletRequest(portletRequest);
-
-		ThemeDisplay themeDisplay = (ThemeDisplay)portletRequest.getAttribute(
-			WebKeys.THEME_DISPLAY);
-
-		InputStream inputStream = null;
-
-		try {
-			String parameterName = getParameterName();
-
-			String fileName = uploadPortletRequest.getFileName(parameterName);
-			String contentType = uploadPortletRequest.getContentType(
-				parameterName);
-			long size = uploadPortletRequest.getSize(parameterName);
-
-			validateFile(fileName, size);
-
-			inputStream = uploadPortletRequest.getFileAsStream(parameterName);
-
-			String uniqueFileName = getUniqueFileName(
-				themeDisplay, fileName);
-
-			return addFileEntry(
-				themeDisplay, uniqueFileName, inputStream, contentType);
-		}
-		catch (IOException ioe) {
-			throw new SystemException(ioe);
-		}
-		finally {
-			StreamUtil.cleanUp(inputStream);
-		}
-	}
-
-	protected abstract FileEntry addFileEntry(
-			ThemeDisplay themeDisplay, String fileName, InputStream inputStream,
-			String contentType)
-		throws PortalException;
+public abstract class BaseBlogsImageUploadHandler
+	extends BaseUniqueFileNameUploadHandler {
 
 	@Override
 	protected void checkPermission(
@@ -115,45 +66,12 @@ public abstract class BaseBlogsImageUploadHandler extends BaseUploadHandler {
 		}
 	}
 
-	protected abstract FileEntry fetchFileEntry(
-			ThemeDisplay themeDisplay, String fileName)
-		throws PortalException;
-
 	protected long getMaxFileSize() {
 		return PrefsPropsUtil.getLong(PropsKeys.BLOGS_IMAGE_MAX_SIZE);
 	}
 
 	protected String getParameterName() {
 		return "imageSelectorFileName";
-	}
-
-	protected String getUniqueFileName(
-			ThemeDisplay themeDisplay, String fileName)
-		throws PortalException {
-
-		FileEntry fileEntry = fetchFileEntry(themeDisplay, fileName);
-
-		if (fileEntry == null) {
-			return fileName;
-		}
-
-		int suffix = 1;
-
-		for (int i = 0; i < _UNIQUE_FILE_NAME_TRIES; i++) {
-			String curFileName = FileUtil.appendParentheticalSuffix(
-				fileName, String.valueOf(suffix));
-
-			fileEntry = fetchFileEntry(themeDisplay, curFileName);
-
-			if (fileEntry == null) {
-				return curFileName;
-			}
-
-			suffix++;
-		}
-
-		throw new PortalException(
-			"Unable to get a unique file name for " + fileName);
 	}
 
 	@Override
@@ -239,7 +157,5 @@ public abstract class BaseBlogsImageUploadHandler extends BaseUploadHandler {
 		throw new BlogImageNameException(
 			"Invalid image for file name " + fileName);
 	}
-
-	private static final int _UNIQUE_FILE_NAME_TRIES = 50;
 
 }
